Fixet remark in client

+++ Client/ClientSocket.h
@@ -3,7 +3,7 @@
 class ClientSocket
 {
 public:
-	virtual bool Connect(const std::string& host_name, const unsigned int port) = 0;//connect to host
+	virtual bool Connect(const std::string& host_name, const unsigned short port) = 0;//connect to host
 	virtual bool Disconnect()=0;//disconnect to host, cloce socket
 	virtual bool Send(const std::string& message)=0;//send information
 };
